Replace GOES-16 cloudmask with GOES-East + CLAVR-X algorithm
Found a different data source:
http://cimss.ssec.wisc.edu/clavrx/google_earth/images/clavrx_goes13_YYYY_zzz_HHmm/cmask.png

Updates every 15 minutes, and it's already in EPSG:4326. Some artefacts of
the reprojection near the poles, but this is vastly more convenient than
the HDF, which maps lonlats to raster pixels; I don't even know how I would
project the sky intercept lonlats onto the image.

Note that it takes the KML at
http://cimss.ssec.wisc.edu/clavrx/google_earth/goes_east_kml/goes_east.html ,
not the image. The KML is needed for georeferencing.